organized files into pages & components
also updated all imports
