+++ pages/careerQuiz/edu-background.tsx
@@ -1,8 +1,8 @@
 import { useRouter } from "next/router";
 import React from "react";
+import ProgressBar from "../../components/quizzes/progressbar";
+import QuizNavbar from "../../components/quizzes/quiznavbar";
 import { Button } from "../../components/ui/Button";
-import ProgressBar from "../studentPortal/quizcomponent/progressbar";
-import QuizNavbar from "../studentPortal/quizcomponent/quiznavbar";
 const EducationBackground = () => {
   const router = useRouter();
